+++ PacManGame/EnemyRed.cpp
@@ -22,7 +22,7 @@ void EnemyRed::actor( ) {
 	int key = -1;
 	for ( int i = 0; i < 4; i++ ) {
 		Vector next_pos = getPos( ) + DIR[ i ] * MapParameter::CHIP_SIZE;
-		int root_num = field->getRootNum( target.x / MapParameter::CHIP_SIZE, target.y / MapParameter::CHIP_SIZE, next_pos.x / MapParameter::CHIP_SIZE, next_pos.y  / MapParameter::CHIP_SIZE );
+		int root_num = field->getRootNum( ( int )( target.x / MapParameter::CHIP_SIZE ), ( int )( target.y / MapParameter::CHIP_SIZE ), ( int )( next_pos.x / MapParameter::CHIP_SIZE ), ( int )( next_pos.y  / MapParameter::CHIP_SIZE ) );
 		if ( min_root > root_num && root_num >= 0 ) {
 			key = i;
 			min_root = root_num;
